# Misc scene-related bugfixes and optimizations