--- a/Code/Tools/SceneAPI/SceneCore/Utilities/PatternMatcher.h
+++ b/Code/Tools/SceneAPI/SceneCore/Utilities/PatternMatcher.h
@@ -45,7 +45,8 @@ namespace AZ
                 SCENE_CORE_API PatternMatcher(const char* pattern, MatchApproach matcher);
                 SCENE_CORE_API PatternMatcher(const AZStd::string& pattern, MatchApproach matcher);
                 SCENE_CORE_API PatternMatcher(AZStd::string&& pattern, MatchApproach matcher);
-                SCENE_CORE_API PatternMatcher(const PatternMatcher& rhs) = default;
+                SCENE_CORE_API PatternMatcher(AZStd::span<AZStd::string_view> patterns, MatchApproach matcher);
+                SCENE_CORE_API PatternMatcher(const PatternMatcher& rhs);
                 SCENE_CORE_API PatternMatcher(PatternMatcher&& rhs);
 
                 SCENE_CORE_API PatternMatcher& operator=(const PatternMatcher& rhs);
@@ -54,16 +55,16 @@ namespace AZ
                 SCENE_CORE_API bool LoadFromJson(rapidjson::Document::ConstMemberIterator member);
 
                 SCENE_CORE_API bool MatchesPattern(const char* name, size_t nameLength) const;
-                SCENE_CORE_API bool MatchesPattern(const AZStd::string& name) const;
+                SCENE_CORE_API bool MatchesPattern(AZStd::string_view name) const;
 
                 SCENE_CORE_API const AZStd::string& GetPattern() const;
                 SCENE_CORE_API MatchApproach GetMatchApproach() const;
 
                 static void Reflect(AZ::ReflectContext* context);
             private:
-                AZStd::string m_pattern;
+                AZStd::vector<AZStd::string> m_patterns;
                 MatchApproach m_matcher = MatchApproach::PostFix;
-                mutable AZStd::unique_ptr<AZStd::regex> m_regexMatcher;
+                mutable AZStd::vector<AZStd::unique_ptr<AZStd::regex>> m_regexMatchers;
             };
         } // SceneCore
     } // SceneAPI