--- a/Gems/Atom/RPI/Code/Source/RPI.Builders/Model/ModelAssetBuilderComponent.cpp
+++ b/Gems/Atom/RPI/Code/Source/RPI.Builders/Model/ModelAssetBuilderComponent.cpp
@@ -172,8 +172,6 @@ namespace AZ
                 m_systemInputAssemblyBufferPoolId = assetIdOutcome.GetValue();
             }
 
-            m_createdSubId.clear();
-
             m_modelName = context.m_group.GetName();
 
             const auto& scene = context.m_scene;