… Gems/SceneProcessing/Registry/SoftNameSettings.setreg