## What does this PR do?

This is a set of bugfixes and optimizations that I discovered while trying to test FBX Settings for a file with a lot of submeshes:
* Optimization (PatternMatcher.cpp): changed Scene's PatternMatcher to only create the regex pattern matcher once on first use. This changed the time to display the LECT.fbx FBX Settings (which contains 27 submeshes) from 3.5 seconds to 2.99 seconds. The savings scale by the number of submeshes and the number of regex patterns.
* Optimization (NodeListSelectionWidget.cpp): changed the method of populating the widget's combobox list. The previous method worked fine for a few entries, but seemed to get exponentially worse as the number of entries went up. The new version seems to have a consistently negligible performance cost.
* Optimization (ManifestWidgetPage.cpp): changed ManifestWidgetPage to avoid redundant ScrollToBottom() queueing. This changed the time to display the LECT.fbx FBX Settings (which contains 27 submeshes) from 2.99 seconds to 2.05 seconds. The savings scale by the number of submeshes.
* Optimization (PatternMatcher.cpp): changed PatternMatcher to contain an array of patterns, and changed the soft name patterns to use an array of Postfix matches instead of a regex. This changed the time to display the LECT.fbx FBX Settings (which contains 27 submeshes) from 2.05 seconds to 1.86 seconds.
* Bugfix (MeshFeatureProcessor.cpp): Asset Browser thumbnail rendering would intermittently assert and crash due to a MeshLoader instance that went invalid in a queued lambda between queueing and execution. This was fixed by making the instance a shared_ptr.
* Bugfix (ModelAssetBuilderComponent.cpp): the sub ID list shouldn't be cleared between mesh groups, or else it won't accurately catch collisions. This causes sub ID collisions to get significantly earlier in the export process, causing the AP failure to get reported anywhere from seconds to minutes sooner.

## How was this PR tested?

* Used PIX to capture timings for each of the optimizations
* Verified for the bugfixes that the suspected problem cases were actually occurring with temporary asserts, then implemented the solutions, then verified that the solutions worked through multiple Editor runs and testing known-broken FBX files.
* Created a few new unit tests for PatternMatcher.